feat: refactor tag types and consolidate tag action item definitions
